working on file fixing requests. routes and controller methods created, still need more things done in documentHelper
