+++ app/models/document.rb
@@ -8,16 +8,6 @@ class Document < ApplicationRecord
 	  # File is now an instance variable so it can be
 	  # accessed in the validation.
 	  @file = params.delete(:file)
-	  @sort_by_input = params.delete(:sort_by)
-	  @rmv_duplicate_input = params.delete(:rmv_duplicate)
-	  @word_occurrence_input = params.delete(:word_occurrence)
-	  @customize_input = params.delete(:customize)
-	  @request_hash = {"sort_by" => @sort_by_input, "rmv_duplicate" => @rmv_duplicate_input, "word_occurrence" => @word_occurrence_input, "customize" => @customize_input}
-
-	  puts "@sort_by = '#{@sort_by_input}' and is it truthy? #{@sort_by_input ? "true" : "false"} and class is: #{@sort_by_input.class}"
-	  puts "@customize = '#{@customize_input}' and is it truthy? #{@customize_input ? "true" : "false"} and class is: #{@customize_input.class}"
-
-	  fix_the_file(@request_hash)
 	  super
 	  if @file
 	    self.name = sanitize_filename(@file.original_filename)
@@ -25,28 +15,6 @@ class Document < ApplicationRecord
 	    self.original_file = @file.read.force_encoding('Windows-1252').encode('UTF-8')
 	    self.content_status = "Pending"
 	  end
-	end
-
-	def fix_the_file hash_in
-		hash_in.delete_if {|key, value| !value}
-		hash_in.each {|key, value| self.public_send(key) if self.respond_to? key}
-		puts "here's the new hash_in: #{hash_in}"
-	end
-
-	def sort_by
-		puts "sort_by method called"
-	end
-
-	def rmv_duplicate
-		puts "rmv method called"
-	end
-
-	def word_occurrence
-		puts "word_occurrence method called"
-	end
-
-	def customize
-		puts "customize method called"
 	end
 
 	NUM_BYTES_IN_MEGABYTE = 1048576
